requirements.txt
psycopg[binary]==3.2.10
psycopg-pool==3.3.0
APScheduler==3.10.4
pytz==2024.1
requests==2.31.0
finvizfinance==0.14.7
python-dotenv==1.0.0
anthropic==0.84.0
twilio==9.4.6
flask==3.1.0
gspread==6.1.4
google-auth==2.36.0
pytest==8.4.2
